backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
pytest==8.3.4
pytest-django==4.8.0
coverage==7.6.12
django-filter==24.1
requests==2.32.3 